pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "echo-mcp-https-server"
version = "0.1.0"
description = "Simple HTTPS MCP Server with Echo Tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mcp>=1.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
echo-mcp-server = "echo_mcp_server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["echo_mcp_server"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39" 